pyproject.toml
[project]
name = "automating-excersise-routine-udemy-day-49"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "selenium>=4.43.0",
]
